+++ src/core/custom-text-input/custom-text-input.tsx
@@ -12,6 +12,7 @@ interface Props {
   returnKeyType?: ReturnKeyType;
   returnKeyLabel?: string;
   errorMessage?: string;
+  blurOnSubmit?: boolean;
   autoCapitalize?: 'characters' | 'none' | 'sentences' | 'words';
   onChange: (value: string) => void;
   onSubmit?: () => void;
@@ -19,7 +20,7 @@ interface Props {
 
 export const CustomTextInput = ({
   value, state, maxLength, autoCapitalize, placeholder, returnKeyType, returnKeyLabel,
-  errorMessage, keyboardType, children,
+  errorMessage, keyboardType, children, blurOnSubmit,
   onChange, onSubmit,
 }: Props) => {
   return (
@@ -35,6 +36,7 @@ export const CustomTextInput = ({
         keyboardType={keyboardType}
         returnKeyType={returnKeyType}
         returnKeyLabel={returnKeyLabel}
+        blurOnSubmit={blurOnSubmit}
       />
       {children ?? null}
       {(state === false && errorMessage) ? <CustomTextInputErrorText children={errorMessage} /> : null}