+++ src/dictionarly/components/dictionarly-end-modal/dictionarly-end-modal.tsx
@@ -5,15 +5,16 @@ import { useLocalize } from '@core/hooks/use-localize.hook'
 import { CustomButton } from '@core/custom-button/custom-button'
 import { COLOR } from '@core/colors/colors.constants'
 import {
-  DictionarlyEndModalButtonsContainer, DictionarlyEndModalContainer, DictionarlyModalText,
+  DictionarlyEndModalButtonsContainer, DictionarlyEndModalContainer, DictionarlyModalLabel, DictionarlyModalText,
 } from './dictionarly-end-modal.styled'
 
 interface Props {
   modalizeRef: React.MutableRefObject<Modalize | null>;
   state: boolean | null;
+  word: string;
 }
 
-export const DictionarlyEndModal = ({ modalizeRef, state }: Props) => {
+export const DictionarlyEndModal = ({ modalizeRef, state, word }: Props) => {
   const localize = useLocalize()
   const navigation = useNavigation<any>()
 
@@ -21,6 +22,7 @@ export const DictionarlyEndModal = ({ modalizeRef, state }: Props) => {
     <Modalize ref={modalizeRef} adjustToContentHeight>
       <DictionarlyEndModalContainer>
         <DictionarlyModalText children={state ? localize().success : localize().failed} />
+        <DictionarlyModalLabel state={state} children={word} />
 
         <DictionarlyEndModalButtonsContainer>
           <CustomButton color={COLOR.DODGER_BLUE} onPress={navigation.goBack} title={localize().try_again} />